test: migrate part of DatePicker tests
